+++ app/models/concerns/cartodb_central_synchronizable.rb
@@ -42,7 +42,6 @@ module Concerns
           :geocoding_block_price, :map_view_block_price,
           :twitter_datasource_enabled, :twitter_datasource_block_size,
           :twitter_datasource_block_price, :twitter_datasource_quota,
-          :here_maps_enabled, :stamen_maps_enabled, :rainbow_maps_enabled,
           :new_dashboard_enabled]
         when :update
           [:seats, :quota_in_bytes, :display_name, :description, :website,
@@ -50,7 +49,6 @@ module Concerns
           :geocoding_block_price, :map_view_block_price,
           :twitter_datasource_enabled, :twitter_datasource_block_size,
           :twitter_datasource_block_price, :twitter_datasource_quota,
-          :here_maps_enabled, :stamen_maps_enabled, :rainbow_maps_enabled,
           :new_dashboard_enabled]
         end
       elsif self.is_a?(User)
@@ -64,8 +62,8 @@ module Concerns
         :twitter_datasource_enabled, :twitter_datasource_block_size,
         :twitter_datasource_block_price, :twitter_datasource_quota,
         :soft_twitter_datasource_limit,
+        :google_sign_in,
         :arcgis_datasource_enabled,
-        :here_maps_enabled, :stamen_maps_enabled, :rainbow_maps_enabled,
         :new_dashboard_enabled, :private_maps_enabled]
       end
     end
@@ -88,7 +86,7 @@ module Concerns
           :sync_tables_enabled, :table_quota, :twitter_username, :upgraded_at,
           :user_timeout, :username, :website, :soft_geocoding_limit,
           :twitter_datasource_enabled, :soft_twitter_datasource_limit,
-          :arcgis_datasource_enabled
+          :arcgis_datasource_enabled, :google_sign_in
         )
         case action
         when :create
@@ -118,7 +116,7 @@ module Concerns
     end
 
     def sync_data_with_cartodb_central?
-      Cartodb.config[:cartodb_central_api].present? && Cartodb.config[:cartodb_central_api]['username'].present? && Cartodb.config[:cartodb_central_api]['password'].present?
+      Cartodb::Central.sync_data_with_cartodb_central?
     end
 
     def cartodb_central_client