+++ NEWS.md
@@ -1,5 +1,8 @@
 3.8.1 (2015-01-xx)
 ------------------
+* Added config for basemaps [#1954], see
+  [documentation](https://github.com/CartoDB/cartodb/wiki/How-to-configure-basemaps-in-CartoDB)
+* Added oEmbed support for visualizations [#1965](https://github.com/CartoDB/cartodb/issues/1965)
 
 3.8.0 (2015-01-30)
 ------------------